requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai-whisper==20231117
yt-dlp==2024.10.22
python-multipart==0.0.6
requests>=2.32.2
